requirements.txt
python-mpd2>=3.0.5
Pillow>=9.2.0
ez-setup>=0.9
aiostream>=0.4.5